+++ src/test/java/org/opensearch/knn/index/AdvancedFilteringUseCasesIT.java
@@ -22,6 +22,7 @@ import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.knn.KNNRestTestCase;
 import org.opensearch.knn.NestedKnnDocBuilder;
 import org.opensearch.knn.index.engine.KNNEngine;
+import org.opensearch.knn.common.annotation.ExpectRemoteBuildValidation;
 
 import java.io.IOException;
 import java.util.List;
@@ -103,6 +104,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      * }
      */
     @SneakyThrows
+    @ExpectRemoteBuildValidation
     public void testFiltering_whenNestedKNNAndFilterFieldWithNestedQueries_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields
@@ -169,6 +171,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      * }
      */
     @SneakyThrows
+    @ExpectRemoteBuildValidation
     public void testFiltering_whenNestedKNNAndFilterFieldWithNoNestedContextInFilterQuery_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields
@@ -232,6 +235,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      *
      */
     @SneakyThrows
+    @ExpectRemoteBuildValidation
     public void testFiltering_whenNestedKNNAndNonNestedFilterFieldWithNonNestedFilterQuery_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields
@@ -299,6 +303,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      * }
      */
     @SneakyThrows
+    @ExpectRemoteBuildValidation
     public void testFiltering_whenNonNestedKNNAndNestedFilterFieldWithNestedFilterQuery_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields
@@ -386,6 +391,7 @@ public class AdvancedFilteringUseCasesIT extends KNNRestTestCase {
      * }
      */
     @SneakyThrows
+    @ExpectRemoteBuildValidation
     public void testFiltering_whenNonNestedKNNAndNestedFilterAndNonNestedFieldWithNestedAndNonNestedFilterQuery_thenSuccess() {
         for (final String engine : enginesToTest) {
             // Set up the index with nested k-nn and metadata fields