+++ src/test/java/org/opensearch/knn/index/KNNESSettingsTestIT.java
@@ -16,6 +16,7 @@ import org.opensearch.knn.KNNRestTestCase;
 import org.opensearch.knn.KNNResult;
 import org.opensearch.knn.index.query.KNNQueryBuilder;
 import org.opensearch.knn.plugin.stats.StatNames;
+import org.opensearch.knn.common.annotation.ExpectRemoteBuildValidation;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -35,6 +36,7 @@ public class KNNESSettingsTestIT extends KNNRestTestCase {
 
     public static final int ALWAYS_BUILD_GRAPH = 0;
 
+    @ExpectRemoteBuildValidation
     public void testKNNLegacySpaceTypeIndexingTest() throws IOException, ParseException {
         // Configure space_type at index level. This is deprecated and will be removed in the future.
         final Settings indexSettings = Settings.builder().put("index.knn", true).put("knn.algo_param.ef_search", 100).build();